+++ src/cn/yapeteam/yolbi/render/shader/Shader.java
@@ -1,18 +1,18 @@
 package cn.yapeteam.yolbi.render.shader;
 
 public abstract class Shader {
-    public int x, y, width, height;
+    public float x, y, width, height;
     public int identifier = -1;
     public int level = 1;
     public boolean antiAlias = false;
 
     public abstract int dispose(float relativeX, float relativeY, float screenWidth, float screenHeight, int pixel);
 
-    public int getRealWidth() {
+    public float getRealWidth() {
         return width / level;
     }
 
-    public int getRealHeight() {
+    public float getRealHeight() {
         return height / level;
     }
 }
